@@ -17,8 +17,8 @@
     <resteasy.version>3.15.1.Final</resteasy.version>
     <!--<version.server.bom>20.0.0.Final</version.server.bom>-->
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
 		<java.version>11</java.version>
   </properties>
 
@@ -77,6 +77,11 @@
           <version>0.11.2</version>
           <scope>runtime</scope>
         </dependency>
+        <dependency>
+          <groupId>com.fasterxml.jackson.core</groupId>
+          <artifactId>jackson-databind</artifactId>
+          <version>2.11.1</version>
+      </dependency>
     </dependencies>
   
     <build>
