@@ -1,6 +1,6 @@
 package it.soseng.unibo.airlineService.repository;
 
-import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -15,9 +15,9 @@ import it.soseng.unibo.airlineService.model.FlightOffer;
  */
 public interface FlightOfferRepository extends JpaRepository<FlightOffer, Long> {
 
-    @Query(value = "SELECT * FROM FLIGHT_OFFERS o WHERE o.DEPARTURE = ?1 AND O.ARRIVAL = ?2 AND o.DEPARTURE_DATE_TIME = ?3 " +
+    @Query(value = "SELECT * FROM FLIGHT_OFFERS o WHERE o.DEPARTURE_AIRPORT_ADDRESS = ?1 AND O.ARRIVAL_AIRPORT_ADDRESS = ?2 AND o.DEPARTURE_DATE_TIME = ?3 " +
                         "AND o.ARRIVAL_DATE_TIME = ?4", nativeQuery = true)
-    public ArrayList<FlightOffer> searchFlightOffers(String departure, String arrival, LocalDateTime departureDate, 
-                                                        LocalDateTime arrivalDate);
+    public ArrayList<FlightOffer> searchFlightOffers(String departure, String arrival, ZonedDateTime departureDate, 
+                                                        ZonedDateTime arrivalDate);
 
 }
